move layout definitions to new file
